refactor: divided interfactes from service.ts into services/ service and validation
