+++ src/di/domain.ts
@@ -1,7 +1,7 @@
 import UserRepository from '@domain/user/repository/UserRepository'
 import { userCreateValidation } from '@domain/user/services/helpers/userValidation'
 import { IRepositoryUser, IUser } from '@interfaces/domain/repository'
-import { IServiceValidationGroup } from '@interfaces/domain/service'
+import { IServiceValidationGroup } from '@interfaces/domain/services/validation'
 import { container } from 'tsyringe'
 
 container.registerSingleton<IRepositoryUser>(
