+++ src/domain/AbstractService.ts
@@ -1,5 +1,5 @@
 import { IEndPointsRepository } from 'interfaces/domain/repository'
-import { IEndPointsService } from 'interfaces/domain/service'
+import { IEndPointsService } from 'interfaces/domain/services/service'
 
 export default abstract class AbstractService implements IEndPointsService {
   repository: IEndPointsRepository
