+++ src/domain/user/services/helpers/userValidation.ts
@@ -1,5 +1,5 @@
 import { IUser } from '@interfaces/domain/repository'
-import { IServiceValidationGroup } from '@interfaces/domain/service'
+import { IServiceValidationGroup } from '@interfaces/domain/services/validation'
 import userCreateSchema from './userSchemas'
 
 export const userCreateValidation: IServiceValidationGroup<IUser> = {
